// packages/libs/registry-sdk/src/lib/schemas/repermitApp.ts
import { EXAMPLE_WALLET_ADDRESS } from '../constants';
import { z } from './openApiZod';

export const repermitAppRequest = z
  .object({
    userControllerAddress: z
      .string()
      .regex(/^0x[a-fA-F0-9]{40}$/)
      .openapi({
        description: 'EOA address that controls the user smart wallet',
        example: EXAMPLE_WALLET_ADDRESS,
      }),
  })
  .strict();

export const repermitAppResponse = z
  .object({
    repermitDataToSign: z.object({}).passthrough().openapi({
      description:
        'The EIP2771 TypedData to sign that will repermit the app in the Vincent contracts',
    }),
  })
  .strict();

export const completeRepermitRequest = z
  .object({
    typedDataSignature: z.string().openapi({
      description: 'The signature of the typed data from the user wallet',
      example: '0x...',
    }),
    repermitDataToSign: z.object({}).passthrough().openapi({
      description:
        'The repermitDataToSign object returned from repermit-app. Generated by the Gelato SDK.',
    }),
  })
  .strict();

export const completeRepermitResponse = z
  .object({
    transactionHash: z.string().openapi({
      description: 'The transaction hash of the completed repermit transaction',
      example: '0x9d260d1bbe075be0cda52a3271df062748f3182ede91b3aae5cd115f7b26552b',
    }),
  })
  .strict();