// packages/libs/registry-sdk/src/lib/schemas/unpermitApp.ts
import { EXAMPLE_WALLET_ADDRESS } from '../constants';
import { z } from './openApiZod';

export const unpermitAppRequest = z
  .object({
    appVersion: z.number().int().positive().openapi({
      description: 'The version of the app to unpermit',
      example: 1,
    }),
    userControllerAddress: z
      .string()
      .regex(/^0x[a-fA-F0-9]{40}$/)
      .openapi({
        description: 'EOA address that controls the user smart wallet',
        example: EXAMPLE_WALLET_ADDRESS,
      }),
  })
  .strict();

export const unpermitAppResponse = z
  .object({
    unpermitDataToSign: z.object({}).passthrough().openapi({
      description:
        'The EIP2771 TypedData to sign that will unpermit the app from the Vincent contracts',
    }),
  })
  .strict();

export const completeUnpermitRequest = z
  .object({
    typedDataSignature: z.string().openapi({
      description: 'The signature of the typed data from the user wallet',
      example: '0x...',
    }),
    unpermitDataToSign: z.object({}).passthrough().openapi({
      description:
        'The unpermitDataToSign object returned from unpermit-app. Generated by the Gelato SDK.',
    }),
  })
  .strict();

export const completeUnpermitResponse = z
  .object({
    transactionHash: z.string().openapi({
      description: 'The transaction hash of the completed unpermit transaction',
      example: '0x9d260d1bbe075be0cda52a3271df062748f3182ede91b3aae5cd115f7b26552b',
    }),
  })
  .strict();